requirements.txt
Flask==0.12
Jinja2==2.9.5
PyYAML==3.12
markdown2==2.3.3
feedgen==0.5.1
